move internal/feature to src/pkg
